+++ controllers/index.js
@@ -1,5 +1,26 @@
-export { default as authController } from "./authController.js";
-export { default as helpController } from "./helpController.js";
-export { default as boardsController } from "./boardsController.js";
-export { default as columnsController } from "./columnsController.js";
-export { default as cardsController } from "./cardsController.js";
+export { default as changeDisplay } from "./auth/changeDisplay.js";
+export { default as changeTheme } from "./auth/changeTheme.js";
+export { default as getCurrent } from "./auth/getCurrentUser.js";
+export { default as login } from "./auth/login.js";
+export { default as logout } from "./auth/logout.js";
+export { default as register } from "./auth/registration.js";
+export { default as updateUser } from "./auth/updateUser.js";
+
+export { default as helpEmail } from "./email/helpEmail.js";
+
+export { default as addBoard } from "./boards/addBoard.js";
+export { default as deleteBoard } from "./boards/deleteBoard.js";
+export { default as getAllBoards } from "./boards/getAllBoards.js";
+export { default as getOneBoard } from "./boards/getOneBoard.js";
+export { default as updateBoard } from "./boards/updateBoard.js";
+
+export { default as addColumn } from "./columns/addColumn.js";
+export { default as deleteColumn } from "./columns/deleteColumn.js";
+export { default as getColumn } from "./columns/getColumn.js";
+export { default as updateColumn } from "./columns/updateColumn.js";
+
+export { default as addCard } from "./cards/addCard.js";
+export { default as deleteCard } from "./cards/deleteCard.js";
+export { default as getCard } from "./cards/getCard.js";
+export { default as moveCard } from "./cards/moveCard.js";
+export { default as updateCard } from "./cards/updateCard.js";
