+++ helpers/getBackground.js
