+++ middlewars/handleMongooseError.js
@@ -1,0 +1,7 @@
+const handleMongooseError = (error, data, next) => {
+  const { code, name } = error;
+  error.status = code === 11000 && name === "MongoServerError" ? 409 : 400;
+  next();
+};
+
+export default handleMongooseError;
