--- middlewars/hooks.js
@@ -1,10 +1,0 @@
-export const handleMongooseError = (error, data, next) => {
-  const { code, name } = error;
-  error.status = code === 11000 && name === "MongoServerError" ? 409 : 400;
-  next();
-};
-
-export const validateAtUpdate = function (next) {
-  this.options.runValidators = true;
-  next();
-};
