+++ middlewars/index.js
@@ -1,10 +1,6 @@
-export { default as HttpError } from "./HttpError.js";
+export { default as handleMongooseError } from "./handleMongooseError.js";
+export { default as validateAtUpdate } from "./validateAtUpdate.js";
+export { default as isValidId } from "./isValidId.js";
 export { default as isEmptyBody } from "./isEmptyBody.js";
-export { handleMongooseError } from "./hooks.js";
-export { validateAtUpdate } from "./hooks.js";
 export { default as authenticate } from "./authenticate.js";
 export { default as uploadCloud } from "./uploadCloud.js";
-export { default as uploadAvatar } from "./uploadAvatar.js";
-export { default as sendEmail } from "./sendEmail.js";
-export { default as isValidId } from "./isValidId.js";
-export { default as getBackground } from "./getBackground.js";
