+++ middlewars/isEmptyBody.js
@@ -1,4 +1,4 @@
-import { HttpError } from "./index.js";
+import { HttpError } from "../helpers/index.js";
 
 const isEmptyBody = (req, res, next) => {
   const { length } = Object.keys(req.body);
