+++ middlewars/validateAtUpdate.js
@@ -1,0 +1,6 @@
+const validateAtUpdate = function (next) {
+  this.options.runValidators = true;
+  next();
+};
+
+export default validateAtUpdate;
